+++ tech1-framework-b2b-mongodb-security-jwt/src/main/java/io/tech1/framework/b2b/mongodb/security/jwt/events/publishers/SecurityJwtIncidentPublisher.java
@@ -5,7 +5,6 @@ import io.tech1.framework.incidents.domain.registration.IncidentRegistration1;
 import io.tech1.framework.incidents.domain.registration.IncidentRegistration1Failure;
 import io.tech1.framework.incidents.domain.session.IncidentSessionExpired;
 import io.tech1.framework.incidents.domain.session.IncidentSessionRefreshed;
-import org.springframework.scheduling.annotation.Async;
 
 public interface SecurityJwtIncidentPublisher {
     void publishAuthenticationLogin(IncidentAuthenticationLogin incident);