+++ tech1-framework-b2b-mongodb-security-jwt/src/main/java/io/tech1/framework/b2b/mongodb/security/jwt/repositories/InvitationCodeRepository.java
@@ -3,6 +3,7 @@ package io.tech1.framework.b2b.mongodb.security.jwt.repositories;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.db.DbInvitationCode;
 import io.tech1.framework.domain.base.Username;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.data.mongodb.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -12,6 +13,9 @@ import static io.tech1.framework.domain.utilities.exceptions.ExceptionsMessagesU
 
 @Repository
 public interface InvitationCodeRepository extends MongoRepository<DbInvitationCode, String> {
+    // ================================================================================================================
+    // Spring Data
+    // ================================================================================================================
     List<DbInvitationCode> findByOwner(Username username);
     List<DbInvitationCode> findByInvitedIsNull();
     DbInvitationCode findByValue(String value);
@@ -25,4 +29,19 @@ public interface InvitationCodeRepository extends MongoRepository<DbInvitationCo
         assertNonNullOrThrow(invitationCode, entityNotFound("DbInvitationCode", invitationCodeId));
         return invitationCode;
     }
+
+    // ================================================================================================================
+    // Queries
+    // ================================================================================================================
+    @Query(value = "{ 'invited': { '$exists': true}}")
+    List<DbInvitationCode> findByInvitedAlreadyUsed();
+
+    @Query(value = "{ 'invited': { '$exists': false}}")
+    List<DbInvitationCode> findByInvitedNotUsed();
+
+    @Query(value = "{ 'invited': { '$exists': true}}", delete = true)
+    void deleteByInvitedAlreadyUsed();
+
+    @Query(value = "{ 'invited': { '$exists': false}}", delete = true)
+    void deleteByInvitedNotUsed();
 }