+++ tech1-framework-b2b-mongodb-security-jwt/src/main/java/io/tech1/framework/b2b/mongodb/security/jwt/repositories/UserRepository.java
@@ -4,15 +4,64 @@ import io.tech1.framework.b2b.mongodb.security.jwt.domain.db.DbUser;
 import io.tech1.framework.domain.base.Email;
 import io.tech1.framework.domain.base.Username;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.data.mongodb.repository.Query;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Set;
 
+import static io.tech1.framework.domain.base.AbstractAuthority.SUPER_ADMIN;
+
 @Repository
 public interface UserRepository extends MongoRepository<DbUser, String> {
+    // ================================================================================================================
+    // Constants
+    // ================================================================================================================
+    SimpleGrantedAuthority SUPERADMIN = new SimpleGrantedAuthority(SUPER_ADMIN);
+
+    // ================================================================================================================
+    // Spring Data
+    // ================================================================================================================
     DbUser findByEmail(Email email);
     DbUser findByUsername(Username username);
     List<DbUser> findByUsernameIn(Set<Username> usernames);
     List<DbUser> findByUsernameIn(List<Username> usernames);
+
+    // ================================================================================================================
+    // Queries
+    // ================================================================================================================
+    @Query(value = "{ 'authorities': ?0}")
+    List<DbUser> findByAuthority(SimpleGrantedAuthority authority);
+
+    @Query(value = "{ 'authorities': ?0}", fields = "{ 'id': 0, 'username' : 1}")
+    List<DbUser> findByAuthorityProjectionUsernames(SimpleGrantedAuthority authority);
+
+    @Query(value = "{ 'authorities': { '$ne': ?0}}")
+    List<DbUser> findByAuthorityNotEqual(SimpleGrantedAuthority authority);
+
+    @Query(value = "{ 'authorities': { '$ne': ?0}}", fields = "{ 'id': 0, 'username' : 1}")
+    List<DbUser> findByAuthorityNotEqualProjectionUsernames(SimpleGrantedAuthority authority);
+
+    @Query(value = "{ 'authorities': ?0}", delete = true)
+    void deleteByAuthority(SimpleGrantedAuthority authority);
+
+    @Query(value = "{ 'authorities': { '$ne': ?0}}", delete = true)
+    void deleteByAuthorityNotEqual(SimpleGrantedAuthority authority);
+
+    default List<DbUser> findByAuthoritySuperadmin() {
+        return this.findByAuthority(SUPERADMIN);
+    }
+
+    default List<DbUser> findByAuthorityNotSuperadmin() {
+        return this.findByAuthorityNotEqual(SUPERADMIN);
+    }
+
+    default void deleteByAuthoritySuperadmin() {
+        this.deleteByAuthority(SUPERADMIN);
+    }
+
+    default void deleteByAuthorityNotSuperadmin() {
+        this.deleteByAuthorityNotEqual(SUPERADMIN);
+    }
 }