+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/reflections/ReflectionProperty.java
@@ -2,6 +2,7 @@ package io.tech1.framework.domain.reflections;
 
 import lombok.Data;
 
+import java.time.ZoneId;
 import java.util.Arrays;
 
 import static io.tech1.framework.domain.asserts.Asserts.assertNonNullOrThrow;
@@ -29,15 +30,32 @@ public class ReflectionProperty {
         this.propertyName = propertyName;
         this.propertyValue = propertyValue;
 
+        // WARNING: supports only String[] and ZoneId (on 5+ cases refactoring or extraction required)
         var isArray = nonNull(this.propertyValue) && this.propertyValue.getClass().isArray();
-
-        // WARNING: draft support only array String[]
+        boolean isArrayOfStrings;
         if (isArray) {
+            var array = (Object[]) this.propertyValue;
+            isArrayOfStrings = array[0] instanceof String;
+        } else {
+            isArrayOfStrings = false;
+        }
+        var isZoneId = nonNull(this.propertyValue) && this.propertyValue instanceof ZoneId;
+
+        if (isArrayOfStrings) {
+            var castedPropertyValue = (String[]) this.propertyValue;
             this.readableValue = String.format(
                     "%s.%s: `%s`",
                     uncapitalize(parentPropertyName),
                     uncapitalize(this.propertyName),
-                    Arrays.toString((String[]) this.propertyValue)
+                    Arrays.toString(castedPropertyValue)
+            );
+        } else if (isZoneId) {
+            var castedPropertyValue = (ZoneId) this.propertyValue;
+            this.readableValue = String.format(
+                    "%s.%s: `%s`",
+                    uncapitalize(parentPropertyName),
+                    uncapitalize(this.propertyName),
+                    castedPropertyValue.getId()
             );
         } else {
             this.readableValue = String.format(