+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/tuples/TuplePresence.java
@@ -1,0 +1,18 @@
+package io.tech1.framework.domain.tuples;
+
+import lombok.Data;
+
+// Lombok
+@Data(staticConstructor = "of")
+public class TuplePresence<A> {
+    private final boolean present;
+    private final A value;
+
+    public static <A> TuplePresence<A> present(A value) {
+        return TuplePresence.of(true, value);
+    }
+
+    public static <A> TuplePresence<A> absent(A value) {
+        return TuplePresence.of(false, value);
+    }
+}